# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  owner: 'owner',
  staff: 'staff',
} as const;

export type MovementType = typeof MovementType[keyof typeof MovementType];


export const MovementType = {
  stock_in: 'stock_in',
  used_in_production: 'used_in_production',
  sold: 'sold',
  damaged: 'damaged',
  missing_lost: 'missing_lost',
  returned: 'returned',
} as const;

export type UnitType = typeof UnitType[keyof typeof UnitType];


export const UnitType = {
  kg: 'kg',
  bags: 'bags',
  liters: 'liters',
  boxes: 'boxes',
  pieces: 'pieces',
  trays: 'trays',
} as const;

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  daily: 'daily',
  weekly: 'weekly',
  missing: 'missing',
  damaged: 'damaged',
  branch_activity: 'branch_activity',
} as const;

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  role: UserRole;
  branchId?: number | null;
  branchName?: string | null;
  createdAt: string;
}

export interface SyncUserRequest {
  clerkId: string;
  name: string;
  email: string;
}

export interface CreateUserRequest {
  name: string;
  email: string;
  role: UserRole;
  branchId?: number | null;
}

export interface UpdateUserRequest {
  name?: string;
  role?: UserRole;
  branchId?: number | null;
}

export interface Branch {
  id: number;
  name: string;
  city: string;
  manager: string;
  phone: string;
  createdAt: string;
}

export interface CreateBranchRequest {
  name: string;
  city: string;
  manager: string;
  phone: string;
}

export interface UpdateBranchRequest {
  name?: string;
  city?: string;
  manager?: string;
  phone?: string;
}

export interface BranchStats {
  branchId: number;
  branchName: string;
  totalItems: number;
  lowStockItems: number;
  totalMovementsToday: number;
  missingLostToday: number;
  damagedToday: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit: UnitType;
  minThreshold: number;
  branchId: number;
  branchName?: string | null;
  isLowStock: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateInventoryItemRequest {
  name: string;
  category: string;
  quantity: number;
  unit: UnitType;
  minThreshold: number;
  branchId: number;
}

export interface UpdateInventoryItemRequest {
  name?: string;
  category?: string;
  quantity?: number;
  unit?: UnitType;
  minThreshold?: number;
  branchId?: number;
}

export interface StockMovement {
  id: number;
  itemId: number;
  itemName?: string | null;
  branchId: number;
  branchName?: string | null;
  userId: number;
  userName?: string | null;
  type: MovementType;
  quantity: number;
  note?: string | null;
  createdAt: string;
}

export interface CreateStockMovementRequest {
  itemId: number;
  branchId: number;
  type: MovementType;
  quantity: number;
  note?: string | null;
}

export interface AuditLog {
  id: number;
  userId: number;
  userName?: string | null;
  branchId: number;
  branchName?: string | null;
  itemId: number;
  itemName?: string | null;
  quantityChange: number;
  movementType: MovementType;
  note?: string | null;
  timestamp: string;
}

export interface LowStockItem {
  id: number;
  name: string;
  quantity: number;
  minThreshold: number;
  unit: UnitType;
  branchId: number;
  branchName?: string | null;
}

export interface TopUsedItem {
  itemId: number;
  itemName: string;
  totalUsed: number;
  unit: UnitType;
}

export interface BranchOverview {
  branchId: number;
  branchName: string;
  totalItems: number;
  lowStockCount: number;
  movementsToday: number;
}

export interface RecentMovement {
  id: number;
  itemName: string;
  branchName: string;
  userName: string;
  type: MovementType;
  quantity: number;
  unit: UnitType;
  createdAt: string;
}

export interface DashboardSummary {
  totalItems: number;
  totalLowStock: number;
  totalMissingToday: number;
  totalDamagedToday: number;
  branchOverviews: BranchOverview[];
  recentMovements: RecentMovement[];
  lowStockAlerts: LowStockItem[];
  topUsedItems: TopUsedItem[];
}

export type ListInventoryItemsParams = {
branchId?: number;
category?: string;
search?: string;
lowStock?: boolean;
};

export type ListStockMovementsParams = {
branchId?: number;
itemId?: number;
type?: MovementType;
dateFrom?: string;
dateTo?: string;
limit?: number;
};

export type ListAuditLogsParams = {
branchId?: number;
userId?: number;
itemId?: number;
movementType?: MovementType;
dateFrom?: string;
dateTo?: string;
limit?: number;
};

export type GetDashboardSummaryParams = {
branchId?: number;
};

export type DownloadReportParams = {
type: ReportType;
format: DownloadReportFormat;
branchId?: number;
dateFrom?: string;
dateTo?: string;
};

export type DownloadReportFormat = typeof DownloadReportFormat[keyof typeof DownloadReportFormat];


export const DownloadReportFormat = {
  csv: 'csv',
  pdf: 'pdf',
} as const;