# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BakeryStock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/sync:
    post:
      operationId: syncUser
      tags: [auth]
      summary: Sync Clerk user to local DB on first login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUserRequest"
      responses:
        "200":
          description: User synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List all staff users (owner only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create a staff account (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [auth]
      summary: Get user by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [auth]
      summary: Update user (owner only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete user (owner only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchRequest"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch (owner only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBranchRequest"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch (owner only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /branches/{id}/stats:
    get:
      operationId: getBranchStats
      tags: [branches]
      summary: Get inventory stats for a branch
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Branch stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchStats"
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemRequest"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    UserRole:
      type: string
      enum: [owner, staff]
...
    UnitType:
      type: string
      enum: [kg, bags, liters, boxes, pieces, trays]
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: integer
          nullable: true
        branchName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - name
        - email
        - role
        - createdAt
    SyncUserRequest:
      type: object
      properties:
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
      required:
        - clerkId
        - name
        - email
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: integer
          nullable: true
      required:
        - name
        - email
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: integer
          nullable: true
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        manager:
          type: string
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - city
        - manager
        - phone
        - createdAt
    CreateBranchRequest:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        manager:
          type: string
        phone:
          type: string
      required:
        - name
        - city
        - manager
        - phone
    UpdateBranchRequest:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        manager:
          type: string
        phone:
          type: string
    BranchStats:
      type: object
      properties:
        branchId:
          type: integer
        branchName:
          type: string
        totalItems:
          type: integer
        lowStockItems:
          type: integer
        totalMovementsToday:
          type: integer
        missingLostToday:
          type: number
        damagedToday:
          type: number
      required:
        - branchId
        - branchName
        - totalItems
        - lowStockItems
        - totalMovementsToday
        - missingLostToday
        - damagedToday
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          $ref: "#/components/schemas/UnitType"
        minThreshold:
          type: number
        branchId:
          type: integer
        branchName:
          type: string
          nullable: true
        isLowStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - quantity
        - unit
        - minThreshold
        - branchId
        - isLowStock
        - createdAt
        - updatedAt
    CreateInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          $ref: "#/components/schemas/UnitType"
        minThreshold:
          type: number
        branchId:
          type: integer
      required:
        - name
        - category
        - quantity
        - unit
        - minThreshold
        - branchId
    UpdateInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          $ref: "#/components/schemas/UnitType"
        minThreshold:
          type: number
        branchId:
          type: integer
...