# lib/api-zod/src/generated/types/createBranchRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBranchRequest {
  name: string;
  city: string;
  manager: string;
  phone: string;
}