# lib/api-zod/src/generated/types/createInventoryItemRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UnitType } from './unitType';

export interface CreateInventoryItemRequest {
  name: string;
  category: string;
  quantity: number;
  unit: UnitType;
  minThreshold: number;
  branchId: number;
}