# lib/api-zod/src/generated/types/syncUserRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SyncUserRequest {
  clerkId: string;
  name: string;
  email: string;
}