# lib/api-zod/src/generated/types/unitType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */

export type UnitType = typeof UnitType[keyof typeof UnitType];


export const UnitType = {
  kg: 'kg',
  bags: 'bags',
  liters: 'liters',
  boxes: 'boxes',
  pieces: 'pieces',
  trays: 'trays',
} as const;